requirements.txt
fastapi
requests
pydantic
pydantic-settings
python-dotenv
python-multipart
uvicorn
httpx
ag-ui-protocol